Segmentation/Task 3/code/requirements.txt
Pillow==8.4.0
torch 
torchvision 
matplotlib==3.7.1
torchmetrics
tensorboard
opencv-python
lmdb
argparse
easydict
PyYAML